chore: Move RedissionRedisInstance to own file
